+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/Item` styling to match the latest GUI
 - `limestone/KeyGuide` styling to match the latest GUI
 - `limestone/RadioItem` styling to match the latest GUI
+- `limestone/Scroller` focus area to be separated from the vertical and horizontal scrollbars when `focusableScrollbar` is `"byEnter"`
 - `limestone/Switch` styling to match the latest GUI
 
 ## [1.10.0] - 2026-05-08